feat(gruntfile): add bowerRelease task to publish library on bower
I've created a new repository, http://github.com/caitp/angular-drop-bower, where the bower release
for this project will live. Updating it is automated via the `grunt release` task, so that should
hopefully be nice and easy.

All that remains is to add the new installation instructions to the demo page and website, horray.

Closes #43